refactor: 예측률 조회 기능 Querydsl로 구현 및 db 설정 파일 추가 (#51)

+++ backend/src/main/java/com/idk/api/user/dto/MyPageResponse.java
@@ -38,10 +38,10 @@ public class MyPageResponse {
     @AllArgsConstructor(access = AccessLevel.PRIVATE)
     public static class Rate{
         private Long id;
-        private int ballotCount;
-        private int correctCount;
+        private long ballotCount;
+        private long correctCount;
 
-        public static Rate build(Long id, int ballotCount, int correctCount){
+        public static Rate build(Long id, long ballotCount, long correctCount){
             return Rate.builder()
                     .id(id)
                     .ballotCount(ballotCount)
